requirements.txt
streamlitpython-docxpdfminer.sixopenaipandasscikit-learn